+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## v5.2.6 - 2017-03-25
+### Added:
+ - HttpClient `dontUseProxy` method to stop sending requests via proxy
+
 ### Fixed:
  - Additional info added in pins *related* method.